Attack Scene
Added attack scene with a script, created a debug enemy for the debug purposes, deleted the killzone from the slime since it`s doesnt serve anything.

Player now is able to attack with the visible hitbox(red) and kill the enemy.

Various enemies(mobs) can be added to "enemies" group for further works.